requirements.txt
altair==4.2.0
asgiref==3.5.2
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
blinker==1.5
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
decorator==5.1.1
Django==4.1
django-admin-list-filter-dropdown==1.0.3
djangorestframework==3.13.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.0.0
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
importlib-metadata==4.12.0
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.14.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
numpy==1.23.2
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
protobuf==3.20.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pydeck==0.8.0b1
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rich==12.5.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
sqlparse==0.4.2
stack-data==0.5.0
streamlit==1.12.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
zipp==3.8.1

pyodbc~=4.0.39